теперь действительно генерируется extern, Исправлены предупреждения с индексами int стал size_t в CPPClassUnit, CSClassUnit, JAVAClassUnit. убраны CONST и VIRTUAL из enum Java

+++ CPPclassunit.h
@@ -10,7 +10,7 @@ public:
 
     void add( const std::shared_ptr< Unit >& unit, Flags flags ) override
     {
-        int accessModifier = PRIVATE;
+        size_t accessModifier = PRIVATE;
         if( flags < ACCESS_MODIFIERS.size() ) {
             accessModifier = flags;
         }
